requirements.txt
python-telegram-bot==20.7
websockets==12.0
pandas==2.1.4
pandas-ta==0.3.14b
numpy==1.26.3
scipy==1.12.0
python-dotenv==1.0.0
aiohttp==3.9.1